… browser/staticcheck.conf
… cmd/data-usage-cache_gen.go
… cmd/data-usage-cache_gen_test.go
… cmd/data-usage_test.go
… cmd/fastwalk.go
… cmd/fs-v1-multipart_test.go
… cmd/fs-v1_test.go
… cmd/gateway/azure/gateway-azure.go
… cmd/gateway/s3/gateway-s3-sse.go
… cmd/generic-handlers.go
… cmd/global-heal.go
… cmd/http-stats.go
… cmd/iam-object-store.go
… cmd/jwt/parser_test.go
… cmd/logger/target/http/http.go
… cmd/naughty-disk_test.go
… cmd/object-api-datatypes.go
… cmd/object-api-errors.go
… cmd/object-api-interface.go
… cmd/object-api-utils.go
… cmd/object-handlers.go
… cmd/posix-diskid-check.go
… cmd/posix.go
… cmd/server-main.go
… cmd/server-startup-msg.go
… cmd/service.go
… cmd/signals.go
… cmd/storage-interface.go
… cmd/storage-rest-client.go
… cmd/storage-rest-common.go